+++ src/services/helpers/gantt/color/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./startConfidenceColor";
+export * from "./startConfidenceColor";
