+++ src/services/helpers/gantt/ganttCell/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./GanttCell";
+export * from "./GanttCell";
